No need call toInt

+++ core/src/main/scala/org/apache/spark/util/io/ChunkedByteBuffer.scala
@@ -60,7 +60,7 @@ private[spark] class ChunkedByteBuffer(var chunks: Array[ByteBuffer]) {
     for (bytes <- getChunks()) {
       val capacity = bytes.limit()
       while (bytes.position() < capacity && bytes.remaining() > 0) {
-        val ioSize = Math.min(bytes.remaining(), NIO_BUFFER_LIMIT.toInt)
+        val ioSize = Math.min(bytes.remaining(), NIO_BUFFER_LIMIT)
         bytes.limit(bytes.position + ioSize)
         channel.write(bytes)
       }
